requirements.txt
einops==0.6.0
matplotlib==3.6.2
numpy==1.23.4
Pillow==9.3.0
torch==1.13.0+cu116
torchvision==0.14.0+cu116
